feat: show indeterminate progress indicator on deleting and pasting, refactor code
