+++ core/main/src/main/java/com/rk/activities/main/MainContent.kt
@@ -189,7 +189,6 @@ fun MainContent(
 
             if (fileActionDialog != null) {
                 FileActionDialog(
-                    modifier = Modifier,
                     file = fileActionDialog!!,
                     root = (currentTab as? FileTreeTab)?.root,
                     onDismissRequest = { fileActionDialog = null },
