+++ core/main/src/main/java/com/rk/components/FileActionDialog.kt
@@ -1,16 +1,8 @@
 package com.rk.components
 
-import android.content.Context
 import android.content.Intent
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
@@ -23,36 +15,25 @@ import androidx.compose.material.icons.outlined.Refresh
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.PrimaryScrollableTabRow
-import androidx.compose.material3.Tab
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.rk.DefaultScope
 import com.rk.activities.main.MainActivity
 import com.rk.activities.terminal.Terminal
 import com.rk.file.FileObject
+import com.rk.file.FileOperations
 import com.rk.file.FileWrapper
-import com.rk.file.openWith
-import com.rk.file.to_save_file
 import com.rk.filetree.FileTreeTab
 import com.rk.filetree.FileTreeViewModel
 import com.rk.filetree.addProject
@@ -65,7 +46,6 @@ import com.rk.icons.XedIcons
 import com.rk.resources.drawables
 import com.rk.resources.fillPlaceholders
 import com.rk.resources.getFilledString
-import com.rk.resources.getString
 import com.rk.resources.strings
 import com.rk.settings.Settings
 import com.rk.settings.app.InbuiltFeatures
@@ -73,19 +53,11 @@ import com.rk.tabs.editor.EditorTab
 import com.rk.utils.errorDialog
 import com.rk.utils.showTerminalNotice
 import com.rk.utils.toast
-import java.text.DateFormat
-import java.text.NumberFormat
-import java.util.Date
-import java.util.Locale
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import org.apache.commons.net.io.Util.copyStream
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun FileActionDialog(
-    modifier: Modifier = Modifier,
     file: FileObject,
     root: FileObject?,
     onDismissRequest: () -> Unit,
@@ -239,12 +211,15 @@ fun FileActionDialog(
                                 val isCut = FileOperations.isCut
                                 val clipboardFile = FileOperations.clipboard!!
                                 val clipboardParentFile = clipboardFile.getParentFile()
-                                pasteFile(
-                                    context = context,
-                                    sourceFile = clipboardFile,
-                                    destinationFolder = file,
-                                    isCut = isCut,
-                                )
+
+                                fileTreeViewModel.withFileOperation {
+                                    FileOperations.pasteFile(
+                                        context = context,
+                                        sourceFile = clipboardFile,
+                                        destinationFolder = file,
+                                        isCut = isCut,
+                                    )
+                                }
 
                                 if (FileOperations.isCut) {
                                     MainActivity.instance?.apply {
@@ -284,7 +259,7 @@ fun FileActionDialog(
                     // description = stringResource(strings.save_desc),
                     onClick = {
                         // This would typically open a file picker
-                        FileOperations.saveAs(context, file)
+                        FileOperations.saveAs(file)
                         // showXedDialog = true
                         onDismissRequest()
                     },
@@ -390,7 +365,8 @@ fun FileActionDialog(
             fileName = file.getName(),
             onConfirm = {
                 scope.launch {
-                    val success = FileOperations.deleteFile(file)
+                    var success = false
+                    fileTreeViewModel.withFileOperation { success = FileOperations.deleteFile(file) }
 
                     if (!success) {
                         toast(strings.delete_failed)
@@ -514,42 +490,6 @@ fun FileActionDialog(
     }
 }
 
-// File Operations utility class
-object FileOperations {
-    var clipboard: FileObject? = null
-    var isCut: Boolean = false
-
-    suspend fun copyToClipboard(file: FileObject, isCut: Boolean = false) {
-        clipboard = file
-        this.isCut = isCut
-    }
-
-    suspend fun deleteFile(file: FileObject): Boolean {
-        return try {
-            file.delete()
-        } catch (e: Exception) {
-            errorDialog(e)
-            false
-        }
-    }
-
-    suspend fun openWithExternalApp(context: Context, file: FileObject) {
-        openWith(context, file)
-    }
-
-    fun saveAs(context: Context, file: FileObject) {
-        to_save_file = file
-        MainActivity.instance?.fileManager?.requestOpenDirectoryToSaveFile(file)
-    }
-
-    fun addFile(parentFile: FileObject) {
-        val intent = Intent(Intent.ACTION_OPEN_DOCUMENT)
-        intent.addCategory(Intent.CATEGORY_OPENABLE)
-        intent.type = "*/*"
-        MainActivity.instance?.fileManager?.requestAddFile(parentFile)
-    }
-}
-
 @Composable
 fun CloseConfirmationDialog(projectName: String, onConfirm: () -> Unit, onDismiss: () -> Unit) {
     AlertDialog(
@@ -579,287 +519,3 @@ fun DeleteConfirmationDialog(fileName: String, onConfirm: () -> Unit, onDismiss:
         dismissButton = { TextButton(onClick = onDismiss) { Text(stringResource(strings.cancel)) } },
     )
 }
-
-data class ContentProgress(val totalSize: Long, val totalItems: Long)
-
-private suspend fun calculateContent(folder: FileObject, onProgress: (ContentProgress) -> Unit = {}) {
-    var totalSize = 0L
-    var totalItems = 0L
-
-    val stack = ArrayDeque<FileObject>()
-    stack.add(folder)
-
-    while (stack.isNotEmpty()) {
-        val current = stack.removeLast()
-        if (current.isDirectory()) {
-            stack.addAll(current.listFiles())
-            totalItems++
-            onProgress(ContentProgress(totalSize, totalItems))
-        } else {
-            totalSize += current.length()
-            totalItems++
-            onProgress(ContentProgress(totalSize, totalItems))
-        }
-    }
-}
-
-enum class PropertyRoutes(val label: String, val route: String) {
-    GENERAL(strings.general.getString(), "general"),
-    ADVANCED(strings.advanced.getString(), "advanced"),
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun PropertiesDialog(file: FileObject, onDismiss: () -> Unit) {
-    val pagerState = rememberPagerState(initialPage = 0) { PropertyRoutes.entries.size }
-    val scope = rememberCoroutineScope()
-
-    val scrollStates = PropertyRoutes.entries.map { rememberScrollState() }
-    val currentScrollState = scrollStates[pagerState.currentPage]
-    val showHorizontalDivider by
-        remember(pagerState.currentPage) { derivedStateOf { currentScrollState.canScrollForward } }
-
-    XedDialog(onDismissRequest = onDismiss) {
-        Column(modifier = Modifier.fillMaxWidth().fillMaxHeight(0.8f)) {
-            Text(
-                text = stringResource(strings.properties),
-                style = MaterialTheme.typography.headlineSmall,
-                fontWeight = FontWeight.Medium,
-                modifier = Modifier.padding(start = 24.dp, top = 24.dp, end = 24.dp, bottom = 16.dp),
-            )
-
-            PrimaryScrollableTabRow(
-                selectedTabIndex = pagerState.currentPage,
-                containerColor = MaterialTheme.colorScheme.surfaceContainer,
-            ) {
-                PropertyRoutes.entries.forEachIndexed { index, destination ->
-                    Tab(
-                        selected = pagerState.currentPage == index,
-                        onClick = { scope.launch { pagerState.animateScrollToPage(index) } },
-                        text = { Text(text = destination.label, maxLines = 2, overflow = TextOverflow.Ellipsis) },
-                    )
-                }
-            }
-
-            HorizontalPager(
-                state = pagerState,
-                beyondViewportPageCount = PropertyRoutes.entries.size,
-                pageSpacing = 16.dp,
-                modifier = Modifier.fillMaxWidth().weight(1f),
-            ) { page ->
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(8.dp),
-                    modifier = Modifier.fillMaxSize().verticalScroll(scrollStates[page]).padding(all = 24.dp),
-                ) {
-                    when (PropertyRoutes.entries[page]) {
-                        PropertyRoutes.GENERAL -> GeneralProperties(file)
-                        PropertyRoutes.ADVANCED -> AdvancedProperties(file)
-                    }
-                }
-            }
-
-            if (showHorizontalDivider) HorizontalDivider()
-
-            Box(modifier = Modifier.padding(start = 24.dp, bottom = 24.dp, end = 24.dp).align(Alignment.End)) {
-                TextButton(onClick = onDismiss) { Text(text = stringResource(strings.close)) }
-            }
-        }
-    }
-}
-
-@Composable
-fun GeneralProperties(file: FileObject) {
-    var size by remember { mutableStateOf(formatFileSize(0)) }
-    var itemsCount by remember { mutableStateOf("0") }
-    val numberFormatter = remember { NumberFormat.getNumberInstance() }
-
-    val lastModified =
-        DateFormat.getDateTimeInstance(DateFormat.MEDIUM, DateFormat.MEDIUM, Locale.getDefault())
-            .format(Date(file.lastModified()))
-
-    LaunchedEffect(file) {
-        if (file.isFile()) {
-            size = formatFileSize(file.length())
-        } else {
-            calculateContent(file) {
-                size = formatFileSize(it.totalSize)
-                itemsCount = numberFormatter.format(it.totalItems)
-            }
-        }
-    }
-
-    InfoRow(stringResource(strings.name), file.getName())
-    InfoRow(stringResource(strings.path), file.getAbsolutePath())
-    InfoRow(
-        label = stringResource(strings.type),
-        value =
-            if (file.isDirectory()) {
-                stringResource(strings.folder)
-            } else {
-                stringResource(strings.file)
-            },
-    )
-    if (file.isDirectory()) {
-        InfoRow(
-            label = stringResource(strings.content),
-            value = stringResource(strings.content_property).fillPlaceholders(itemsCount, size),
-        )
-    } else {
-        InfoRow(label = stringResource(strings.size), value = size)
-    }
-    InfoRow(stringResource(strings.last_modified), lastModified)
-}
-
-@Composable
-fun AdvancedProperties(file: FileObject) {
-    //    PermissionRow(canRead = file.canRead(), canWrite = file.canWrite(), canExecute = file.canExecute())
-    InfoRow(stringResource(strings.permissions), getPseudoPermissions(file))
-    if (InbuiltFeatures.debugMode.state.value) {
-        InfoRow(stringResource(strings.wrapper_type), file.javaClass.simpleName)
-    }
-}
-
-@Composable
-fun InfoRow(label: String, value: String) {
-    OutlinedTextField(value = value, onValueChange = {}, label = { Text(label) }, readOnly = true)
-}
-
-fun getPseudoPermissions(file: FileObject): String {
-    val type =
-        when {
-            file.isDirectory() -> "d"
-            file.isSymlink() -> "l"
-            else -> "-"
-        }
-
-    val r = if (file.canRead()) "r" else "-"
-    val w = if (file.canWrite()) "w" else "-"
-    val x = if (file.canExecute()) "x" else "-"
-    return "$type$r$w$x"
-}
-
-// Utility functions
-private fun formatFileSize(bytes: Long): String {
-    if (bytes < 1024) return "$bytes B"
-    val kb = bytes / 1024.0
-    if (kb < 1024) return String.format(Locale.getDefault(), "%.1f KB", kb)
-    val mb = kb / 1024.0
-    if (mb < 1024) return String.format(Locale.getDefault(), "%.1f MB", mb)
-    val gb = mb / 1024.0
-    return String.format(Locale.getDefault(), "%.1f GB", gb)
-}
-
-/**
- * Pastes a file or folder to the destination directory
- *
- * @param context Android context for content resolver operations
- * @param sourceFile The file or folder to copy/move
- * @param destinationFolder The target directory
- * @param isCut Whether to move (cut) or copy the file
- * @param onProgress Optional callback for progress updates (current file being processed)
- * @return Result indicating success or failure with error message
- */
-suspend fun pasteFile(
-    context: Context,
-    sourceFile: FileObject,
-    destinationFolder: FileObject,
-    isCut: Boolean = false,
-    onProgress: ((String) -> Unit)? = null,
-): Result<Unit> =
-    withContext(Dispatchers.IO) {
-        runCatching {
-            // Validation checks
-            if (!destinationFolder.isDirectory()) {
-                throw IllegalArgumentException("Destination must be a directory")
-            }
-
-            if (!destinationFolder.canWrite()) {
-                throw IllegalStateException("Destination directory is not writable")
-            }
-
-            if (!sourceFile.exists()) {
-                throw IllegalArgumentException("Source file does not exist")
-            }
-
-            // Prevent copying a directory into itself
-            if (sourceFile.isDirectory() && isParentOf(sourceFile, destinationFolder)) {
-                throw IllegalArgumentException("Cannot copy a directory into itself")
-            }
-
-            // Check if target already exists
-            val targetName = sourceFile.getName()
-            if (destinationFolder.hasChild(targetName)) {
-                throw IllegalStateException("A file or folder with name '$targetName' already exists in destination")
-            }
-
-            // Perform the copy operation
-            copyRecursive(context, sourceFile, destinationFolder, onProgress)
-
-            // If it's a cut operation, delete the source
-            if (isCut) {
-                val deleteSuccess = sourceFile.delete()
-                if (!deleteSuccess) {
-                    throw IllegalStateException("Failed to delete source file after moving")
-                }
-            }
-        }
-    }
-
-/** Recursively copies a file or directory */
-private suspend fun copyRecursive(
-    context: Context,
-    sourceFile: FileObject,
-    targetParent: FileObject,
-    onProgress: ((String) -> Unit)?,
-) {
-    onProgress?.invoke("Processing: ${sourceFile.getName()}")
-
-    if (sourceFile.isDirectory()) {
-        // Create target directory
-        val targetDir =
-            targetParent.createChild(false, sourceFile.getName())
-                ?: throw IllegalStateException("Failed to create directory: ${sourceFile.getName()}")
-
-        // Copy all children
-        sourceFile.listFiles().forEach { child -> copyRecursive(context, child, targetDir, onProgress) }
-    } else {
-        // Copy file content
-        val targetFile =
-            targetParent.createChild(true, sourceFile.getName())
-                ?: throw IllegalStateException("Failed to create file: ${sourceFile.getName()}")
-
-        context.contentResolver.openInputStream(sourceFile.toUri())?.use { inputStream ->
-            context.contentResolver.openOutputStream(targetFile.toUri())?.use { outputStream ->
-                copyStream(inputStream, outputStream)
-            } ?: throw IllegalStateException("Failed to open output stream for: ${sourceFile.getName()}")
-        } ?: throw IllegalStateException("Failed to open input stream for: ${sourceFile.getName()}")
-    }
-}
-
-/** Checks if parentDir is a parent of childDir (prevents copying directory into itself) */
-private suspend fun isParentOf(parentDir: FileObject, childDir: FileObject): Boolean {
-    var current: FileObject? = childDir
-    while (current != null) {
-        if (current == parentDir) {
-            return true
-        }
-        current = current.getParentFile()
-    }
-    return false
-}
-
-/** Convenience function for copying files */
-suspend fun copyFile(
-    context: Context,
-    sourceFile: FileObject,
-    destinationFolder: FileObject,
-    onProgress: ((String) -> Unit)? = null,
-): Result<Unit> = pasteFile(context, sourceFile, destinationFolder, isCut = false, onProgress)
-
-/** Convenience function for moving files */
-suspend fun moveFile(
-    context: Context,
-    sourceFile: FileObject,
-    destinationFolder: FileObject,
-    onProgress: ((String) -> Unit)? = null,
-): Result<Unit> = pasteFile(context, sourceFile, destinationFolder, isCut = true, onProgress)
